+++ CHANGES.md
@@ -12,6 +12,7 @@
   interface
 - Query: Added a few example questions specific to CrateDB
 - Outline: Renamed `CRATEDB_CONTEXT_URL` to `ABOUT_CONTEXT_URL`
+- Outline: Fixed `llms_txt` currently does not accept newlines in description fields
 
 ## v0.0.3 - 2025-05-10
 - Outline: Refactored the source of truth for the documentation outline